XPathSetting moved to Configuration
