+++ ReasonCodeExample.XPathTools.Tests/Writers/AttributeFilterTests.cs
@@ -1,5 +1,6 @@
 ﻿using System.Xml.Linq;
 using NUnit.Framework;
+using ReasonCodeExample.XPathTools.Configuration;
 using ReasonCodeExample.XPathTools.Writers;
 
 namespace ReasonCodeExample.XPathTools.Tests.Writers
