+++ ReasonCodeExample.XPathTools/Writers/AttributeFilter.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Xml.Linq;
+using ReasonCodeExample.XPathTools.Configuration;
 
 namespace ReasonCodeExample.XPathTools.Writers
 {
