+++ ReasonCodeExample.XPathTools/Writers/DistinctAttributeFilter.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using System.Xml.Linq;
+using ReasonCodeExample.XPathTools.Configuration;
 
 namespace ReasonCodeExample.XPathTools.Writers
 {
